reorganised training form Added validation split options moved training code to own file created an abortController train nav disabled if not using birdnet
