+++ js/utils/abortController.js
@@ -1,0 +1,9 @@
+const EventEmitter = require('events');
+
+class AbortController extends EventEmitter {
+  abort() {
+    this.emit('abort');
+  }
+}
+
+module.exports = new AbortController();
